+++ src/adapter/http/tests/tests/test_data_ingest.rs
@@ -301,6 +301,7 @@ async fn test_data_push_ingest_upload_token_no_initial_source() {
     let upload_token = UploadTokenBase64Json(UploadToken {
         upload_id,
         file_name: String::from("population.json"),
+        owner_account_id: harness.server_account_id().as_multibase().to_string(),
         content_type: Some(MediaType(String::from("application/json"))),
         content_length: FILE_CONTENT.len(),
     });
@@ -388,6 +389,7 @@ async fn test_data_push_ingest_upload_token_with_initial_source() {
     let upload_token = UploadTokenBase64Json(UploadToken {
         upload_id,
         file_name: String::from("population.json"),
+        owner_account_id: harness.server_account_id().as_multibase().to_string(),
         content_type: Some(MediaType(String::from("application/json"))),
         content_length: FILE_CONTENT.len(),
     });
@@ -467,6 +469,7 @@ async fn test_data_push_ingest_upload_content_type_not_specified() {
     let upload_token = UploadTokenBase64Json(UploadToken {
         upload_id,
         file_name: String::from("population.ndjson"),
+        owner_account_id: harness.server_account_id().as_multibase().to_string(),
         content_type: None,
         content_length: FILE_CONTENT.len(),
     });
@@ -554,6 +557,7 @@ async fn test_data_push_ingest_upload_token_actual_file_different_size() {
     let upload_token = UploadTokenBase64Json(UploadToken {
         upload_id,
         file_name: String::from("population.json"),
+        owner_account_id: harness.server_account_id().as_multibase().to_string(),
         content_type: Some(MediaType(String::from("application/json"))),
         content_length: FILE_CONTENT.len() - 17, // Intentionally wrong file size
     });
@@ -617,6 +621,10 @@ impl DataIngestHarness {
         }
     }
 
+    fn server_account_id(&self) -> &AccountID {
+        &(self.server_harness.server_account().id)
+    }
+
     async fn create_population_dataset(&self, with_schema: bool) -> CreateDatasetResult {
         self.server_harness
             .cli_create_dataset_from_snapshot_use_case()
@@ -668,11 +676,7 @@ impl DataIngestHarness {
 
         let upload_dir = cache_dir
             .join("uploads")
-            .join(
-                AccountID::new_seeded_ed25519(SERVER_ACCOUNT_NAME.as_bytes())
-                    .as_multibase()
-                    .to_string(),
-            )
+            .join(self.server_account_id().as_multibase().to_string())
             .join(upload_id);
 
         std::fs::create_dir_all(upload_dir.clone()).unwrap();